fix(llm/customlockscreen/framedImage): rename, document and pixel polish
