+++ apps/ledger-live-mobile/src/screens/CustomImage/PreviewPreEdit.tsx
@@ -31,9 +31,9 @@ import {
 } from "../../components/CustomImage/imageUtils";
 import { ImageFileUri } from "../../components/CustomImage/types";
 import { targetDisplayDimensions } from "./shared";
-import FramedImage, {
+import StaxFramedImage, {
   previewConfig,
-} from "../../components/CustomImage/FramedImage";
+} from "../../components/CustomImage/StaxFramedImage";
 import ImageProcessor, {
   Props as ImageProcessorProps,
   ProcessorPreviewResult,
@@ -315,7 +315,7 @@ const PreviewPreEdit = ({ navigation, route }: NavigationProps) => {
               alignItems="center"
               justifyContent="center"
             >
-              <FramedImage
+              <StaxFramedImage
                 onError={handlePreviewImageError}
                 fadeDuration={0}
                 source={{ uri: processorPreviewImage?.imageBase64DataUri }}
