+++ apps/ledger-live-mobile/src/screens/Settings/Debug/Features/FetchCustomImage.tsx
@@ -25,9 +25,9 @@ import SelectDevice2, {
 import CustomImageDeviceAction from "../../../../components/CustomImageDeviceAction";
 import ResultDataTester from "../../../../components/CustomImage/ResultDataTester";
 import { ProcessorPreviewResult } from "../../../../components/CustomImage/ImageProcessor";
-import FramedImage, {
+import StaxFramedImage, {
   transferConfig,
-} from "../../../../components/CustomImage/FramedImage";
+} from "../../../../components/CustomImage/StaxFramedImage";
 import { NavigationHeaderBackButton } from "../../../../components/NavigationHeaderBackButton";
 import { ReactNavigationHeaderOptions } from "../../../../components/RootNavigator/types/helpers";
 
@@ -208,7 +208,7 @@ export default function DebugFetchCustomImage() {
             />
             {imageSource ? (
               <Flex flexDirection="row" flexGrow={0}>
-                <FramedImage
+                <StaxFramedImage
                   frameConfig={transferConfig}
                   source={imageSource}
                 />
